+++ src/send_body.rs
@@ -3,8 +3,8 @@ use std::io::{self, Read, Stdin};
 use std::net::TcpStream;
 
 use crate::body::{Body, BodyReader};
-use crate::http;
 use crate::util::private::Private;
+use crate::{http, Error};
 
 /// Request body for sending data via POST, PUT and PATCH.
 ///
@@ -18,23 +18,24 @@ use crate::util::private::Private;
 ///
 pub struct SendBody<'a> {
     inner: BodyInner<'a>,
+    size: Option<Result<u64, Error>>,
     ended: bool,
 }
 
 impl<'a> SendBody<'a> {
     /// Creates an empty body.
     pub fn none() -> SendBody<'static> {
-        BodyInner::None.into()
+        (None, BodyInner::None).into()
     }
 
     /// Creates a body from a shared [`Read`] impl.
     pub fn from_reader(reader: &'a mut dyn Read) -> SendBody<'a> {
-        BodyInner::Reader(reader).into()
+        (None, BodyInner::Reader(reader)).into()
     }
 
     /// Creates a body from an owned [`Read`] impl.
     pub fn from_owned_reader(reader: impl Read + 'static) -> SendBody<'static> {
-        BodyInner::OwnedReader(Box::new(reader)).into()
+        (None, BodyInner::OwnedReader(Box::new(reader))).into()
     }
 
     /// Creates a body to send as JSON from any [`Serialize`](serde::ser::Serialize) value.
@@ -43,7 +44,9 @@ impl<'a> SendBody<'a> {
         value: &impl serde::ser::Serialize,
     ) -> Result<SendBody<'static>, crate::Error> {
         let json = serde_json::to_vec_pretty(value)?;
-        Ok(BodyInner::ByteVec(io::Cursor::new(json)).into())
+        let len = json.len() as u64;
+        let body = (Some(len), BodyInner::ByteVec(io::Cursor::new(json))).into();
+        Ok(body)
     }
 
     pub(crate) fn read(&mut self, buf: &mut [u8]) -> io::Result<usize> {
@@ -73,8 +76,38 @@ impl<'a> SendBody<'a> {
         Ok(n)
     }
 
-    pub(crate) fn body_mode(&self) -> BodyMode {
-        self.inner.body_mode()
+    pub(crate) fn body_mode(&mut self) -> Result<BodyMode, Error> {
+        // Lazily surface a potential error now.
+        let size = match self.size {
+            None => None,
+            Some(Ok(v)) => Some(v),
+            Some(Err(_)) => {
+                // unwraps here are ok because we matched exactly this
+                return Err(self.size.take().unwrap().unwrap_err());
+            }
+        };
+
+        match &self.inner {
+            BodyInner::None => return Ok(BodyMode::NoBody),
+            BodyInner::Body(v) => return Ok(v.body_mode()),
+
+            // The others fall through
+            BodyInner::ByteSlice(_) => {}
+            #[cfg(feature = "json")]
+            BodyInner::ByteVec(_) => {}
+            BodyInner::Reader(_) => {}
+            BodyInner::OwnedReader(_) => {}
+        };
+
+        // Any other body mode could be LengthDelimited depending on whether
+        // we have got a size set.
+        let mode = if let Some(size) = size {
+            BodyMode::LengthDelimited(size)
+        } else {
+            BodyMode::Chunked
+        };
+
+        Ok(mode)
     }
 
     /// Turn this `SendBody` into a reader.
@@ -190,6 +223,7 @@ impl<'a> AsSendBody for SendBody<'a> {
                 BodyInner::Body(v) => BodyInner::Reader(v),
                 BodyInner::OwnedReader(v) => BodyInner::Reader(v),
             },
+            size: self.size.take(),
             ended: self.ended,
         }
     }
@@ -205,94 +239,103 @@ pub(crate) enum BodyInner<'a> {
     OwnedReader(Box<dyn Read>),
 }
 
-impl<'a> BodyInner<'a> {
-    pub fn body_mode(&self) -> BodyMode {
-        match self {
-            BodyInner::None => BodyMode::NoBody,
-            BodyInner::ByteSlice(v) => BodyMode::LengthDelimited(v.len() as u64),
-            #[cfg(feature = "json")]
-            BodyInner::ByteVec(v) => BodyMode::LengthDelimited(v.get_ref().len() as u64),
-            BodyInner::Body(v) => v.body_mode(),
-            BodyInner::Reader(_) => BodyMode::Chunked,
-            BodyInner::OwnedReader(_) => BodyMode::Chunked,
-        }
-    }
-}
-
 impl Private for &[u8] {}
 impl AsSendBody for &[u8] {
     fn as_body(&mut self) -> SendBody {
-        BodyInner::ByteSlice(self).into()
+        let inner = BodyInner::ByteSlice(self);
+        (Some(self.len() as u64), inner).into()
     }
 }
 
 impl Private for &str {}
 impl AsSendBody for &str {
     fn as_body(&mut self) -> SendBody {
-        BodyInner::ByteSlice((*self).as_ref()).into()
+        let inner = BodyInner::ByteSlice((*self).as_ref());
+        (Some(self.len() as u64), inner).into()
     }
 }
 
 impl Private for String {}
 impl AsSendBody for String {
     fn as_body(&mut self) -> SendBody {
-        BodyInner::ByteSlice((*self).as_ref()).into()
+        let inner = BodyInner::ByteSlice((*self).as_ref());
+        (Some(self.len() as u64), inner).into()
     }
 }
 
 impl Private for Vec<u8> {}
 impl AsSendBody for Vec<u8> {
     fn as_body(&mut self) -> SendBody {
-        BodyInner::ByteSlice((*self).as_ref()).into()
+        let inner = BodyInner::ByteSlice((*self).as_ref());
+        (Some(self.len() as u64), inner).into()
     }
 }
 
 impl Private for &String {}
 impl AsSendBody for &String {
     fn as_body(&mut self) -> SendBody {
-        BodyInner::ByteSlice((*self).as_ref()).into()
+        let inner = BodyInner::ByteSlice((*self).as_ref());
+        (Some(self.len() as u64), inner).into()
     }
 }
 
 impl Private for &Vec<u8> {}
 impl AsSendBody for &Vec<u8> {
     fn as_body(&mut self) -> SendBody {
-        BodyInner::ByteSlice((*self).as_ref()).into()
+        let inner = BodyInner::ByteSlice((*self).as_ref());
+        (Some(self.len() as u64), inner).into()
     }
 }
 
 impl Private for &File {}
 impl AsSendBody for &File {
     fn as_body(&mut self) -> SendBody {
-        BodyInner::Reader(self).into()
-    }
-}
-
-impl Private for &TcpStream {}
-impl AsSendBody for &TcpStream {
-    fn as_body(&mut self) -> SendBody {
-        BodyInner::Reader(self).into()
+        let size = lazy_file_size(self);
+        SendBody {
+            inner: BodyInner::Reader(self),
+            size: Some(size),
+            ended: false,
+        }
     }
 }
 
 impl Private for File {}
 impl AsSendBody for File {
     fn as_body(&mut self) -> SendBody {
-        BodyInner::Reader(self).into()
+        let size = lazy_file_size(self);
+        SendBody {
+            inner: BodyInner::Reader(self),
+            size: Some(size),
+            ended: false,
+        }
+    }
+}
+
+fn lazy_file_size(file: &File) -> Result<u64, Error> {
+    match file.metadata() {
+        Ok(v) => Ok(v.len()),
+        Err(e) => Err(e.into()),
+    }
+}
+
+impl Private for &TcpStream {}
+impl AsSendBody for &TcpStream {
+    fn as_body(&mut self) -> SendBody {
+        (None, BodyInner::Reader(self)).into()
     }
 }
 
 impl Private for TcpStream {}
 impl AsSendBody for TcpStream {
     fn as_body(&mut self) -> SendBody {
-        BodyInner::Reader(self).into()
+        (None, BodyInner::Reader(self)).into()
     }
 }
 
 impl Private for Stdin {}
 impl AsSendBody for Stdin {
     fn as_body(&mut self) -> SendBody {
-        BodyInner::Reader(self).into()
+        (None, BodyInner::Reader(self)).into()
     }
 }
 
@@ -307,14 +350,15 @@ impl Private for UnixStream {}
 #[cfg(target_family = "unix")]
 impl AsSendBody for UnixStream {
     fn as_body(&mut self) -> SendBody {
-        BodyInner::Reader(self).into()
+        (None, BodyInner::Reader(self)).into()
     }
 }
 
-impl<'a> From<BodyInner<'a>> for SendBody<'a> {
-    fn from(inner: BodyInner<'a>) -> Self {
+impl<'a> From<(Option<u64>, BodyInner<'a>)> for SendBody<'a> {
+    fn from((size, inner): (Option<u64>, BodyInner<'a>)) -> Self {
         SendBody {
             inner,
+            size: size.map(Ok),
             ended: false,
         }
     }
@@ -323,27 +367,30 @@ impl<'a> From<BodyInner<'a>> for SendBody<'a> {
 impl Private for Body {}
 impl AsSendBody for Body {
     fn as_body(&mut self) -> SendBody {
-        BodyInner::Body(Box::new(self.as_reader())).into()
+        let size = self.content_length();
+        (size, BodyInner::Body(Box::new(self.as_reader()))).into()
     }
 }
 
 impl Private for Response<Body> {}
 impl AsSendBody for Response<Body> {
     fn as_body(&mut self) -> SendBody {
-        BodyInner::Body(Box::new(self.body_mut().as_reader())).into()
+        let size = self.body().content_length();
+        (size, BodyInner::Body(Box::new(self.body_mut().as_reader()))).into()
     }
 }
 
 impl<const N: usize> Private for &[u8; N] {}
 impl<const N: usize> AsSendBody for &[u8; N] {
     fn as_body(&mut self) -> SendBody {
-        BodyInner::ByteSlice(self.as_slice()).into()
+        let inner = BodyInner::ByteSlice((*self).as_ref());
+        (Some(self.len() as u64), inner).into()
     }
 }
 
 impl Private for () {}
 impl AsSendBody for () {
     fn as_body(&mut self) -> SendBody {
-        BodyInner::None.into()
+        (None, BodyInner::None).into()
     }
 }