+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+* Send Content-Length for File (#1100)
 * native-tls transport capture and surface underlying errors (#1093)
 * Bump webpki-roots/webpki-root-certs to 1.0.0 (#1089)
 * Bump rustls-platform-verifier to 0.6.0 (#1089)